Allow make to run parallel jobs
See https://bugs.gentoo.org/show_bug.cgi?id=264273

Make prints warnings, if it is called with "-jN" where N is absent or an
integer greater than one. "... jobserver unavailable ..."

Quote "Wormo" 2009-07-27 05:19:09 UTC:
The cause of this warning message is makefiles that recurse into other
directories without using make's preferred syntax,
  $(MAKE) -C subdir
or
  cd subdir && $(MAKE)
In the case of noweb, the rules are using plain "make" instead of
"$(MAKE)". make is unhappy about the non-standard syntax because it
loses [sic.] track of info needed for parallel builds, and that's why it
spits out the warning messages.
End Quote.